[Optimize] Remove task_wait_timeout_ and node_manager_ from fiber_element
In the current implementation, when creating a FiberElement or attaching it to the ElementManager, the FiberElement retains a pointer to the NodeManager initialized in the ElementManager to trigger an erase when it is destructed. Additionally, FiberElement has an int32_t field called task_wait_timeout_, which is only used for unit testing. These two fields do not need to be part of the FiberElement, as they pertain to the element context rather than the individual element.

In this merge request, these two fields are moved to the ElementManager to reduce the memory usage of FiberElement.

AutoSubmit: true

issue:m-0