[Optimize] Thread safety processing of layout-related methods at the platform layer.
Previously, when the platform layer called layout-related methods, there was a lack of thread safety control, which may result in accessing layout-related methods in non-Layout threads.
This commit binds the Looper of the layout thread to the LayoutNodeManager of the platform layer when creating ShadowNode. When calling layout-related methods through LayoutNodeManager later, it will determine whether the current thread is the Layout thread and perform different operations to ensure thread safety.

issue: m-6083642320