[BugFix][Android] Fix unexpected text content omission.
Flex layout may cause sub-nodes to measure multiple times. The first measurement obtains the width of the text, and then the returned width will be used as a limit for the second measurement. When the width is not an integer, there may be a second measurement width that is smaller than the text width obtained from the first measurement, resulting in omission in text that should display completely.

Return the ceiling value of width.

issue:m-6465081574