[Optimize] Introduce WeakRefPtr to optimize synchronous call
Introduce WeakRefPtr which does not strongly reference the internal RefCounted object. It is mainly used in synchronous call scenarios to avoid modifying the reference count and improve efficiency. It can be implicitly converted to the corresponding RefPtr strong reference pointer. Because WeakRefPtr itself cannot be copied, when it needs to be passed to a closure asynchronously, please call strongify() to convert it to a RefPtr strong reference pointer.

issue: m-6533190588

SkipChecks: macro