[Feature][Styling] Add style objects for simple styling apis
StyleObject is an C++ object that contains parsed style properties and value. And has the ability binding to elements and
updates elements binding to it when updating styles in the StyleObject.

In this commit, we add a base StyleObject class. More functionalities will in next few commits.

issue: #840

doc: https://github.com/lynx-family/lynx/issues/840