[Optimize][Text] Add enableTextLayoutCache switch to control whether text layout caching is enabled.
The current text layout cache is a global cache. Through testing, it was found that after deleting the cache, there is about 10% optimization for single-line text layout indicators on Android, and there is also a certain degree of optimization on iOS.

Determine the impact of using caching on Lynx online pages. Add enableTextLayoutCache pageConfig and settings to experimentally determine the positive and negative benefits of enabling caching.

If there is an overall optimization for closing the cache, the cache will be closed by default afterwards.

issue:m-14790012