[BugFix] fix TemplateData crash with un-initialized Map
as `mData` in TemplateData is a normal field, and it's construction may occurs with `Instruction Rescheduling`,
such that `mData` may be assigned with a un-initialized state in one thread, and be used in another thread which
causes JVM crash.

add `volatile` flag to disable `Instruction Rescheduling` here.

issue: f-23978467567
AutoSubmit: true