[BugFix][Element] Fix reset incorrect property in fiber element with RTL direction
In the current implementation, when resetting a CSS property due to a style change, the direction is not taken into account, leading to incorrect CSS property resets in RTL mode.

In this merge request, before resetting a CSS property, the property ID is adjusted based on the current direction value to ensure the correct CSS property is reset.

issue:m-6732682439

AutoSubmit: true