[BugFix] Fix the error of the Actor object in LynxEngineProxy during the Engine reuse process.
When entering the reuse process through LoadTemplate/Update, we will reset the EngineActor of the Shell, which causes the Actor object in EngineProxy to be different from that in the Shell, resulting in the inability to respond to click events.
Therefore, in this MR, we actively reset the Actor of LynxEngineProxy during ReattachEngine to ensure the consistency of the Engine.

issue: m-6745819543
AutoLand:release/3.4, release_3.4_weekly_20250815

# Conflicts:
#	lynx/platform/android/lynx_android/src/main/java/com/lynx/tasm/core/LynxEngineProxy.java