[BugFix][Event] Fix a crash issue caused by EventDispatcher resetting failure.
description: Because the EventDispatcher is destroyed before the root node, it might cause UserData to fail to reset, leading to a dangling pointer crash when an event is triggered. This is modified to reset UserData when the root node is destroyed to ensure the effectiveness of the EventDispatcher.

issue: m-6795477569

NoLanding: develop, release/3.5